requirements.txt
streamlit>=1.20.0
groq>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
